Fix: change global types to exports
