+++ src/utils/proxy-handler.ts
@@ -1,4 +1,5 @@
 import { v4 as uuidv4 } from 'uuid';
+import { AppWindow } from '../types/dev-server';
 
 declare const window: AppWindow;
